# docs/swagger.yaml
basePath: /api
definitions:
...
  routes.CreateDefinitionRequest:
    properties:
      abbreviation:
        type: string
      name:
        type: string
      suffix:
        type: string
    required:
    - abbreviation
    - name
    type: object
  routes.DefinitionResponse:
    properties:
      abbreviation:
        type: string
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      suffix:
        type: string
      updatedAt:
        type: string
    type: object
  routes.DefinitionsListResponse:
    properties:
      definitions:
        items:
          $ref: '#/definitions/routes.DefinitionResponse'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a wallet service API for managing user accounts, assets, and
    transactions
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Wallet Service API
  version: "1.0"
paths:
...
  /definitions:
    get:
      consumes:
      - application/json
      description: Get all asset definitions with optional pagination
      parameters:
      - description: Limit number of results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.DefinitionsListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all definitions
      tags:
      - definitions
    post:
      consumes:
      - application/json
      description: Create a new asset definition
      parameters:
      - description: Definition creation data
        in: body
        name: definition
        required: true
        schema:
          $ref: '#/definitions/routes.CreateDefinitionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/routes.DefinitionResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new definition
      tags:
      - definitions
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"